+++ nae-spring-core-adapter/pom.xml
@@ -103,16 +103,6 @@
             <artifactId>querydsl-core</artifactId>
             <version>${querydsl.version}</version>
         </dependency>
-        <dependency>
-            <groupId>com.querydsl</groupId>
-            <artifactId>querydsl-apt</artifactId>
-            <version>${querydsl.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.querydsl</groupId>
-            <artifactId>querydsl-mongodb</artifactId>
-            <version>${querydsl.version}</version>
-        </dependency>
     </dependencies>
 
     <build>
@@ -134,30 +124,6 @@
                         <goals>
                             <goal>build-info</goal>
                         </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>com.mysema.maven</groupId>
-                <artifactId>apt-maven-plugin</artifactId>
-                <version>1.1.3</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>process</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>target/generated-sources/java</outputDirectory>
-                            <processors>
-                                <processor>
-                                    com.querydsl.apt.QuerydslAnnotationProcessor
-                                </processor>
-                            </processors>
-                            <options>
-                                <querydsl.excludedClasses>groovy.lang.MetaClass</querydsl.excludedClasses>
-                                <querydsl.unknownAsEmbeddable>true</querydsl.unknownAsEmbeddable>
-                            </options>
-                        </configuration>
                     </execution>
                 </executions>
             </plugin>
